♻️ relocate auth page components from `features/` to `components/`
