+++ frontend/apps/app/app/app/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginPage } from '@/features/auths/LoginPage'
+import { LoginPage } from '@/components/LoginPage'
 
 export default async function Page() {
   return <LoginPage />
