+++ frontend/apps/app/components/LoginPage/services/loginByGithub.ts
